Update README to checkout amplify env before push
Thanks for your tutorial, it's fantastic. I found that the queue wasn't showing up with an `amplify push`; I found in a similar tutorial [here](https://cevo.com.au/post/post-2020-10-14-extend-aws-amplify-with-custom-resources/) that I needed to do an `amplify env checkout <env-name>` to let the amplify CLI pick up the new category; once I'd done this it worked for me.